feat #18 구조화 작업중
