+++ model/__init__.py
@@ -1,0 +1,5 @@
+from .movie_dao import MovieDao
+
+__all__ = [
+    'MovieDao',
+]
