+++ service/__init__.py
@@ -1,0 +1,5 @@
+from .movie_service import MovieService
+
+__all__ = [
+    'MovieService',
+]
